pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "kbase_protein_query_module"
version = "2.0.0"
description = "KBase Protein Query Module"
authors = [
    { name = "Your Name", email = "your@email.com" }
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "torch>=1.13.0",
    "transformers>=4.31.0",
    "sentence-transformers>=2.2.0",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scipy>=1.7.3",
    "scikit-learn>=1.0.2",
    "networkx>=2.6.0",
    "faiss-cpu>=1.7.0",
    "annoy>=1.17.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "plotly>=5.0.0",
    "kaleido>=0.2.1",
    "h5py>=3.6.0",
    "zarr>=2.10.3",
    "tables>=3.7.0",
    "pyarrow>=6.0.0",
    "biopython>=1.79",
    "joblib>=1.1.0",
    "tqdm>=4.62.0",
    "pyyaml>=6.0",
]

[tool.setuptools]
package-dir = {"" = "lib"}

[tool.setuptools.packages.find]
where = ["lib"]

[tool.poetry.dependencies]
pytest = "*" 